+++ App/Views/Posts/index.view.php
@@ -42,8 +42,10 @@ use App\Models\Post;
                                 <h2><?php echo $post->getTitle() ?></h2>
                                 <p><?php echo $post->getText() ?></p>
                                 <a class="btn btn-primary" href="?c=posts&a=content&id=<?php echo $post->getId() ?>">Čítať viac →</a>
-                                <a href="?c=posts&a=delete&id=<?php echo $post->getId() ?>" class="btn btn-danger">Zmazat</a>
-                                <a href="?c=posts&a=edit&id=<?php echo $post->getId() ?>" class="btn btn-warning">Upravit</a>
+                                <?php if ($auth->isLogged()) { ?>
+                                    <a href="?c=posts&a=delete&id=<?php echo $post->getId() ?>" class="btn btn-danger">Zmazat</a>
+                                    <a href="?c=posts&a=edit&id=<?php echo $post->getId() ?>" class="btn btn-warning">Upravit</a>
+                                <?php } ?>
                             </div>
                         </div>
                     </div>
@@ -54,10 +56,26 @@ use App\Models\Post;
         </div>
     </div>
 
-    <div class="row py-3">
-        <div class="col-lg-12">
-            <a class="btn btn-success" href="?c=posts&a=create">Pridaj post →</a>
+    <?php if ($auth->isLogged()) { ?>
+        <div class="row py-3">
+            <div class="col-lg-12">
+                <a class="btn btn-success" href="?c=posts&a=create">Pridaj post →</a>
+            </div>
         </div>
-    </div>
+    <?php } ?>
 
 </div>
+
+<!-- pagination -->
+<nav aria-label="Pagination">
+    <hr class="my-0" />
+    <ul class="pagination justify-content-center my-4">
+        <li class="page-item disabled"><a class="page-link" href="#" tabindex="-1" aria-disabled="true">Newer</a></li>
+        <li class="page-item active" aria-current="page"><a class="page-link" href="#!">1</a></li>
+        <li class="page-item"><a class="page-link" href="#!">2</a></li>
+        <li class="page-item"><a class="page-link" href="#!">3</a></li>
+        <li class="page-item disabled"><a class="page-link" href="#!">...</a></li>
+        <li class="page-item"><a class="page-link" href="#!">15</a></li>
+        <li class="page-item"><a class="page-link" href="#!">Older</a></li>
+    </ul>
+</nav>
